HR-Assistant/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
langchain==0.3.26
chromadb==1.0.15
sentence-transformers==5.0.0
pdfplumber==0.11.7
python-docx==1.2.0
python-dotenv==1.0.0
openai==1.90.0
tiktoken==0.9.0 